requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
faiss-cpu
langchain
langchain-google-genai
reportlab
